+++ NEWS.md
@@ -1,6 +1,8 @@
 # 1.5.0-9004
+* Added / exposed function for collapsing strata on IndividualSamplingParameters (CollapseStrata)
 * Added documentation for dealing with incomplete biological records (#170)
 * Made stratification options for ComputePsuSamplingParameters more flexible (#171)
+* Fixed a bug with handling of NA values in AddLengthGroupStoxBiotic
 * Fixed a bug with missing error detection when trying to report estimates over several strata (#173)
 * Fixed a bug which would cause duplicate samplingUnitIds in the SelectionTable when running ComputePsuSamplingParameters with the option "adHocStoxBiotic" (#172)
 * Fixed a bug which cause the menu for ExtendAnalyticalSamplingFrameCoverage to not show options for StratificationVariables